Isolate wafer-map benchmarks in perf package
