+++ packages/nimble-components/src/tests/nimble-intro.stories.mdx
@@ -18,7 +18,10 @@ Refer to the
 contributing to the component library. To add or update component design documentation,
 refer to the [documentation guide](https://github.com/ni/nimble/tree/main/packages/nimble-components/docs/creating-storybook-component-documentation.md).
 
-See the <a href="./example-client-app" target="_blank">example Angular app</a>, <a href="./blazor-client-app/wwwroot" target="_blank">example Blazor app</a> or <a href="./performance/wafer-map" target="_blank">example Performance app</a> using the components!
+See the <a href="./example-client-app" target="_blank">example Angular app</a> or <a href="./blazor-client-app/wwwroot" target="_blank">example Blazor app</a> using the components!
+
+See component performance tests:
+- <a href="./performance/wafer-map" target="_blank">Wafer Map</a> 
 
 ## How to use this site?
