--- packages/performance/landing/index.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<script type="module" src="wafer-map.ts"></script>
-</head>
-<body>
-</body>
-</html>
