+++ packages/performance/tsconfig.json
@@ -14,6 +14,5 @@
         "noUnusedParameters": true,
         "noImplicitReturns": true,
         "skipLibCheck": true
-    },
-    "include": ["landing"]
+    }
 }
