+++ packages/performance/wafer-map/lighthouserc.js
@@ -1,0 +1,14 @@
+module.exports = {
+    ci: {
+        collect: {
+            staticDistDir: '../dist/wafer-map',
+            numberOfRuns: 1,
+            settings: {
+                onlyAudits: ['user-timings']
+            }
+        },
+        upload: {
+            target: 'temporary-public-storage',
+        },
+    },
+};
