+++ packages/performance/wafer-map/vite.config.js
@@ -1,0 +1,10 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { defineConfig } from 'vite';
+
+// eslint-disable-next-line import/no-default-export
+export default defineConfig({
+    base: './',
+    build: {
+        outDir: '../../dist/wafer-map'
+    }
+});
