+++ packages/site/.eleventy.js
@@ -5,7 +5,7 @@ module.exports = eleventyConfig => {
         '../../packages/nimble-components/dist/storybook': 'storybook',
         '../../angular-workspace/dist/example-client-app': 'storybook/example-client-app',
         '../../packages/nimble-blazor/dist/blazor-client-app': 'storybook/blazor-client-app',
-        '../../packages/performance/landing/dist': 'storybook/performance/wafer-map'
+        '../../packages/performance/dist': 'storybook/performance'
     });
     eleventyConfig.setTemplateFormats([
         'md',
